Create Fizz-Buzz.js
Fizz Buzz Interview question: I replaced Fizz Buzz with Tres and Cinco respectively

'Write a program that prints the numbers from 1 to 100. But for multiples of three print “Tres” instead of the number and for the multiples of five print "Cinco". For numbers which are multiples of both three and five print “Trescinco”.'